Add connected to keypair relationship

+++ plugin/tasks.py
@@ -25,7 +25,6 @@ from plugin.gcp.firewall import FirewallRule
 from plugin.gcp.instance import Instance
 from plugin.gcp.network import Network
 from plugin.gcp.keypair import KeyPair
-from plugin import tags
 
 NAME = 'gcp_name'
 
@@ -71,6 +70,8 @@ def add_instance_tag(gcp_config, instance_name, tag, **kwargs):
 @operation
 @throw_cloudify_exceptions
 def remove_instance_tag(gcp_config, instance_name, tag, **kwargs):
+    if not instance_name:
+        return
     gcp_config['network'] = utils.get_gcp_resource_name(gcp_config['network'])
     instance = Instance(gcp_config,
                         ctx.logger,
@@ -80,16 +81,20 @@ def remove_instance_tag(gcp_config, instance_name, tag, **kwargs):
 @operation
 @throw_cloudify_exceptions
 def add_external_ip(gcp_config, instance_name, **kwargs):
+    # check if the instance has no external ips, only one is supported so far
     gcp_config['network'] = utils.get_gcp_resource_name(gcp_config['network'])
     instance = Instance(gcp_config,
                         ctx.logger,
                         instance_name=instance_name)
     instance.add_access_config()
+    set_ip(instance, relationship=True)
 
 
 @operation
 @throw_cloudify_exceptions
 def remove_external_ip(gcp_config, instance_name, **kwargs):
+    if not instance_name:
+        return
     gcp_config['network'] = utils.get_gcp_resource_name(gcp_config['network'])
     instance = Instance(gcp_config,
                         ctx.logger,
@@ -99,12 +104,15 @@ def remove_external_ip(gcp_config, instance_name, **kwargs):
 @operation
 @throw_cloudify_exceptions
 def delete_instance(gcp_config, **kwargs):
-    name = ctx.instance.runtime_properties[NAME]
+    name = ctx.instance.runtime_properties.get(NAME)
+    if not name:
+        return
     instance = Instance(gcp_config,
                         ctx.logger,
                         instance_name=name)
     instance.delete()
-    ctx.instance.runtime_properties.pop(NAME)
+    if ctx.instance.runtime_properties.get(NAME):
+        ctx.instance.runtime_properties.pop(NAME)
 
 
 @operation
@@ -121,13 +129,16 @@ def create_network(gcp_config, network, **kwargs):
 @operation
 @throw_cloudify_exceptions
 def delete_network(gcp_config, **kwargs):
-    network_name = ctx.instance.runtime_properties[NAME]
+    network_name = ctx.instance.runtime_properties.get(NAME)
+    if not network_name:
+        return
     network = Network(gcp_config,
                       ctx.logger,
                       network=network_name)
 
     network.delete()
-    ctx.instance.runtime_properties.pop(NAME)
+    if ctx.instance.runtime_properties.get(NAME):
+        ctx.instance.runtime_properties.pop(NAME)
 
 
 @operation
@@ -148,14 +159,18 @@ def create_firewall_rule(gcp_config, firewall_rule, **kwargs):
 @operation
 @throw_cloudify_exceptions
 def delete_firewall_rule(gcp_config, **kwargs):
+    firewall_name = ctx.instance.runtime_properties.get(NAME)
+    if not firewall_name:
+        return
     network_name = utils.get_gcp_resource_name(gcp_config['network'])
-    firewall = {'name': ctx.instance.runtime_properties[NAME]}
+    firewall = {'name': firewall_name}
     firewall = FirewallRule(gcp_config,
                             ctx.logger,
                             firewall=firewall,
                             network=network_name)
     firewall.delete()
-    ctx.instance.runtime_properties.pop(NAME)
+    if ctx.instance.runtime_properties.get(NAME):
+        ctx.instance.runtime_properties.pop(NAME)
 
 
 @operation
@@ -180,7 +195,30 @@ def create_keypair(gcp_config, user, private_key_path, **kwargs):
                       private_key_path)
     keypair.create_keypair()
     keypair.add_project_ssh_key(user, keypair.public_key)
+    ctx.instance.runtime_properties['gcp_private_key'] = keypair.private_key
     ctx.instance.runtime_properties['gcp_public_key'] = keypair.public_key
+    keypair.save_private_key()
+
+
+@operation
+@throw_cloudify_exceptions
+def save_private_key(gcp_config, private_key, user, private_key_path, **kwargs):
+    keypair = KeyPair(gcp_config,
+                      ctx.logger,
+                      user,
+                      private_key_path)
+    keypair.private_key = private_key
+    keypair.save_private_key()
+
+
+@operation
+@throw_cloudify_exceptions
+def delete_private_key(gcp_config, user, private_key_path, **kwargs):
+    keypair = KeyPair(gcp_config,
+                      ctx.logger,
+                      user,
+                      private_key_path)
+    keypair.remove_private_key()
 
 
 @operation
@@ -192,17 +230,19 @@ def delete_keypair(gcp_config, user, private_key_path, **kwargs):
                       private_key_path)
     keypair.public_key = ctx.instance.runtime_properties['gcp_public_key']
     keypair.remove_project_ssh_key()
-    keypair.remove_private_key()
+    ctx.instance.runtime_properties.pop('gcp_public_key')
+    ctx.instance.runtime_properties.pop('gcp_private_key')
 
 
-def set_ip(instance):
+def set_ip(instance, relationship=False):
     instances = instance.list()
-    item = utils.get_item_from_gcp_response(
-        'name',
-        ctx.instance.runtime_properties[NAME],
-        instances)
-    ctx.instance.runtime_properties['ip'] = \
-        item['networkInterfaces'][0]['networkIP']
+    item = utils.get_item_from_gcp_response('name', instance.name, instances)
+    if relationship:
+        ctx.target.instance.runtime_properties['gcp_resource_id'] = \
+            item['networkInterfaces'][0]['accessConfigs'][0]['natIP']
+    else:
+        ctx.instance.runtime_properties['ip'] = \
+            item['networkInterfaces'][0]['networkIP']
     # only with one default network interface
 
 
@@ -214,9 +254,16 @@ def create_firewall_structure_from_rules(network, rules):
                 'targetTags': []}
 
     for rule in rules:
-        firewall['sourceTags'].extend(rule.get('source_tags', []))
+        source_tags = rule.get('source_tags', [])
+        for tag in source_tags:
+            tag = utils.get_gcp_resource_name(tag)
+            if tag not in firewall['sourceTags']:
+                firewall['sourceTags'].append(tag)
         firewall['allowed'].extend([{'IPProtocol': rule.get('ip_protocol'),
-                                    'port': rule.get('port', [])}])
-        firewall['sourceRanges'].extend(rule.get('cidr_ip', []))
+                                    'ports': [rule.get('port', [])]}])
+        cidr = rule.get('cidr_ip')
+        if cidr and cidr not in firewall['sourceRanges']:
+            firewall['sourceRanges'].append(cidr)
         firewall['targetTags'].extend(rule.get('target_tags', []))
+    ctx.logger.info(str(firewall))
     return firewall
